requirements.txt
torch
opencv-python
boto3
Pillow
python-dotenv

pandas
ipython
torchvision
pyYAML
tqdm
matplotlib
seaborn
gitpython
scipy
